Fix bugs in the radix testing code

+++ src/seminumerical/radix_conv.h
@@ -1,2 +1,2 @@
 signed char *radix_convi(int a, unsigned r, unsigned *l);
-signed char *radix_convf(float a, unsigned r, unsigned *l);
+signed char *radix_convf(double a, unsigned r, unsigned *l);
